packages/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "langchain>=1.0.3",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
